patched the bugs reported by Weixing
git-svn-id: http://svn.code.sf.net/p/sumo/code/branches/v0_8_2@2308 29f149b2-0c63-11e1-8f5c-a347f6a9acd0

+++ sumo/src/microsim/traffic_lights/MSActuatedPhaseDefinition.h
@@ -14,10 +14,10 @@
 class MSActuatedPhaseDefinition : public MSPhaseDefinition{
 public:
     /// The minimum duration of the phase
-    size_t minDuration;
+    int minDuration;
 
     /// The maximum duration of the phase
-    size_t maxDuration;
+    int maxDuration;
 
     /// stores the timestep of the last on-switched of the phase
     MSNet::Time _lastSwitch;
@@ -27,7 +27,7 @@ public:
     MSActuatedPhaseDefinition(size_t durationArg,
         const std::bitset<64> &driveMaskArg, const std::bitset<64> &breakMaskArg,
         const std::bitset<64> &yellowMaskArg,
-        size_t minDurationArg, size_t maxDurationArg)
+        int minDurationArg, int maxDurationArg)
         : MSPhaseDefinition(durationArg, driveMaskArg,
             breakMaskArg, yellowMaskArg),
         minDuration(minDurationArg), maxDuration(maxDurationArg),
